Accepted float as an argument too

+++ ex1/ft_score_analytics.py
@@ -32,7 +32,7 @@ def main() -> None:
 
     for arg in args:
         try:
-            scores.append(int(arg))
+            scores.append(float(arg))
         except ValueError:
             print(f"Invalid parameter: '{arg}'")
